# docs/swagger.yaml
definitions:
...
  follow.PostFollowStudioRequest:
    properties:
      userId:
        type: integer
    required:
    - userId
    type: object
  follow.PostFollowUserRequest:
    properties:
      userId:
        type: integer
    required:
    - userId
    type: object
  follow.PostUnFollowStudioRequest:
    properties:
      userId:
        type: integer
    required:
    - userId
    type: object
  follow.PostUnFollowUserRequest:
    properties:
      userId:
        type: integer
    required:
    - userId
    type: object
...
info:
  contact:
    email: santhosh@bip.so
    name: API Support
    url: https://bip.so
  description: Bip Backend Platform server.
  license:
    name: Apache 2.0
  termsOfService: https://bip.so/terms-of-service/
  title: Bip Backend Platform
paths:
...
  /v1/follow/studio/follow:
    post:
      parameters:
      - description: Follow studio
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/follow.PostFollowStudioRequest'
      responses: {}
      summary: Follow a Studio
      tags:
      - Followings
...
  /v1/follow/studio/unfollow:
    post:
      parameters:
      - description: unFollow User
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/follow.PostUnFollowStudioRequest'
      responses: {}
      summary: Unfollow a Studio
      tags:
      - Followings
  /v1/follow/user/follow:
    post:
      parameters:
      - description: Follow User
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/follow.PostFollowUserRequest'
      responses: {}
      summary: Follow a User
      tags:
      - Followings
...
  /v1/follow/user/unfollow:
    post:
      parameters:
      - description: unFollow User
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/follow.PostUnFollowUserRequest'
      responses: {}
      summary: Unfollow a User
      tags:
      - Followings
...
schemes:
- http
- https
securityDefinitions:
  bearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"